+++ pgcom/commuter.py
@@ -392,7 +392,6 @@ class Commuter:
             table_name: str,
             data: pd.DataFrame,
             schema: str = 'public',
-            columns: Optional[List[str]] = None,
             format_data: bool = False,
             where: Optional[str] = None
     ) -> None:
@@ -405,8 +404,6 @@ class Commuter:
                 DataFrame from where to insert.
             schema:
                 Name of the schema.
-            columns:
-                
             format_data:
                 Reorder columns and adjust dtypes wrt to table metadata
                 from information_schema.
@@ -416,8 +413,10 @@ class Commuter:
                 DELETE command is not executed if not specified.
         """
 
+        df = data.copy()
+
         if format_data:
-            data = self._format_data(data, table_name, schema)
+            df = self._format_data(df, table_name, schema)
 
         table_name = self._table_name(table_name, schema)
 
@@ -429,7 +428,7 @@ class Commuter:
 
                     # DataFrame to buffer
                     s_buf = StringIO()
-                    data.to_csv(s_buf, index=False, header=False)
+                    df.to_csv(s_buf, index=False, header=False)
                     s_buf.seek(0)
 
                     cur.copy_from(
@@ -437,7 +436,7 @@ class Commuter:
                         table=table_name,
                         sep=',',
                         null='',
-                        columns=columns)
+                        columns=df.columns)
 
                 conn.commit()
             except Exception as e:
@@ -476,8 +475,8 @@ class Commuter:
             self,
             table_name: str,
             data: pd.DataFrame,
-            where: str,
-            schema: str = 'public'
+            schema: str = 'public',
+            where: Optional[str] = None
     ) -> pd.DataFrame:
         """Resolve primary key conflicts in DataFrame.
 
@@ -493,11 +492,11 @@ class Commuter:
                 Name of the table.
             data:
                 DataFrame with primary key conflicts.
+            schema:
+                Name of the schema.
             where:
                 User defined `WHERE` clause used when selecting
                 data from `table_name`.
-            schema:
-                Name of the schema.
 
         Returns:
             pd.DataFrame without primary key conflicts.
@@ -511,7 +510,10 @@ class Commuter:
         if len(p_key) > 0:
             table_name = self._table_name(table_name, schema)
 
-            cmd = f"SELECT * FROM {table_name} WHERE {where}"
+            cmd = 'SELECT * FROM ' + table_name
+            if where is not None:
+                cmd += ' WHERE ' + where
+
             table_data = self.select(cmd)
 
             if not table_data.empty:
@@ -532,8 +534,8 @@ class Commuter:
             table_name: str,
             parent_name: str,
             data: pd.DataFrame,
-            where: str,
             schema: str = 'public',
+            where: Optional[str] = None,
             parent_schema: Optional[str] = None
     ) -> pd.DataFrame:
         """Resolve foreign key conflicts in DataFrame.
@@ -552,11 +554,11 @@ class Commuter:
                 Name of the parent table.
             data:
                 DataFrame with foreign key conflicts.
+            schema:
+                Name of the child table schema.
             where:
                 User defined `WHERE` clause used when selecting
                 data from `table_name`.
-            schema:
-                Name of the child table schema.
             parent_schema:
                 Name of the parent table schema.
 
@@ -576,7 +578,10 @@ class Commuter:
         if len(foreign_key) > 0:
             parent_table = self._table_name(parent_name, parent_schema)
 
-            cmd = f"SELECT * FROM {parent_table} WHERE {where}"
+            cmd = 'SELECT * FROM ' + parent_table
+            if where is not None:
+                cmd += ' WHERE ' + where
+
             parent_data = self.select(cmd)
 
             if not parent_data.empty:
@@ -766,24 +771,24 @@ class Commuter:
         """Formatting DataFrame before applying copy_from.
         """
 
-        # select column attributes
+        # names of the table columns from information schema
         table_columns = self._table_columns(table_name, schema)
+
+        # intersection of DataFrame and table columns
+        columns = []  # type: List[str]
 
         # adjust dtypes of DataFrame columns
         for row in table_columns.itertuples():
             column = row.column_name
 
-            if column not in data.columns:
-                continue
+            if column in data.columns:
+                columns += [column]
 
-            if row.data_type in ['smallint', 'integer', 'bigint']:
-                if data[column].dtype == np.float:
-                    data[column] = data[column].astype(int)
+                if row.data_type in ['smallint', 'integer', 'bigint']:
+                    if data[column].dtype == np.float:
+                        data[column] = data[column].astype(int)
 
-        # set columns order according to ordinal position
-        data = data[table_columns['column_name'].to_list()]
-
-        return data
+        return data[columns]
 
     @staticmethod
     def _table_name(