+++ CHANGELOG.rst
@@ -1,6 +1,17 @@
 ChangeLog
 =========
 
+0.1.4 (2019-01-21)
+------------------
+
+* Changed ``where`` argument type in `:func:`~pgcom.commuter.Commuter.resolve_primary_conflicts`
+from positional to optional
+* Changed ``where`` argument type in `:func:`~pgcom.commuter.Commuter.resolve_foreign_conflicts`
+from positional to optional
+* Fixed bug in copying from DataFrame with incomplete set of columns
+(`#3 <https://github.com/viktorsapozhok/pgcom/issues/3>`_)
+* Added new test
+
 0.1.3 (2019-01-19)
 ------------------